+++ build.sh
@@ -4,17 +4,20 @@ set -e
 
 # make directories
 mkdir -p binaries images
+mkdir -p binaries/user
 
 # compile kernel
-make clean && make
+make -C kernel clean && make -C kernel
+
+# compile libc
+make -C benlibc clean && make -C benlibc
+
+# compile user programs
+make -C user/welcome clean && make -C user/welcome
 
 # assemble bootloader
 nasm -f bin -o binaries/bootsect.bin bootloader/bootsect.asm
 nasm -f bin -o binaries/stage2.bin bootloader/stage2.asm
-
-# compile user programs
-nasm -f elf32 -o obj/pwelcome.o user/welcome.asm
-ld -m elf_i386 -Ttext 0x80000 --oformat binary -o binaries/welcome.bin obj/pwelcome.o
 
 # create and format floppy image
 dd if=/dev/zero of=images/benix.img bs=512 count=2880 conv=notrunc
@@ -22,7 +25,7 @@ dd if=/dev/zero of=images/floppy2.img bs=512 count=2880 conv=notrunc
 mkfs.fat -F 12 images/floppy2.img
 # copy files on secondary floppy
 mcopy -i images/floppy2.img user/README.txt ::
-mcopy -i images/floppy2.img binaries/welcome.bin ::
+mcopy -i images/floppy2.img binaries/user/welcome.bin ::
 
 # copy bootloader and kernel on floppy image
 dd if=binaries/bootsect.bin of=images/benix.img seek=0 conv=notrunc