+++ kernel/cpu/interrupts/syscalls.h
@@ -3,9 +3,9 @@
 #include "klib/types.h"
 
 struct SyscallRegisters {
-    u32 gs, fs, es, ds;
-    u32 edi, esi, ebp, esp, ebx, edx, ecx, eax;
+    u32 edi, esi, ebp, esp;
+    u32 ebx, edx, ecx, eax;
 };
 
-void syscall_handler(struct SyscallRegisters* registers);
+void syscall_handler(struct SyscallRegisters* r);
 void syscall_stub(void);