+++ kernel/init/init.h
@@ -1,0 +1,7 @@
+#pragma once
+
+void install_video(void);
+void install_gdt_idt(void);
+void install_drivers(void);
+void install_proc(void);
+void install_tss(void);