+++ run.sh
@@ -8,4 +8,5 @@ chmod +x build.sh
 qemu-system-i386 \
     -drive format=raw,file=images/benix.img,index=0,if=ide \
     -drive format=raw,file=images/floppy2.img,index=1,if=ide \
-    -boot c -m 512M
+    -boot c -m 512M #\
+    #-d int